First release
This kernels are optimized to work on all Debian based distros, only x64 distros not x86(32bits).

Atom is for Intel Atom CPUs obviously, try it to get a better performance.

Intel and AMD are optimized for most of actual CPUs and older but I'll invite you to try it.